requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

# Environment
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Validation
pydantic==2.5.3
email-validator==2.1.0

# Utilities
pytz==2023.3
